refactor: Remove some unnecessary layers
The `ForgeConfiguration` class was created only to be passed as input to
the `Forge` constructor. No code path, including tests, would ever use
it directly, and all code paths now had to execute two statements.
- Create a `ForgeConfiguration` from the config JSON
- Create a `Forge` from that, including dependencies to interact with
  files and UI

So while it may have made sense from a point of abstraction, the `Forge`
is where it's bound to the actual user interaction; it didn't benefit
the code.

Also the _input_ to the `ForgeConfiguration` is really what I consider
the "forge configuration", so it was hijacking the name.