feat: Don't reset command when changing type
While a "set-value" and "add-to-array" are two completely different
command on a fundamental level. But the share the same two properties,
`key`, and `value`. Seen from the view of the user, they are not that
different, and the type is just a property you can change.

This fix makes sure that it behaves as the user expects; and the new
command maintains the properties of the old command.