fix: SettingsTab would should settings from startup
There was a serious bug that when you changed the settings, they _would_
be stored. BUT the settings tab would still only show the original
settings when reopening.

This made it appear as if the changes were not saved. Any new attempt to
modify the settings would then result in the first set of changes being
discarded.

For some reason, reloading the plugin did not work (is this an Obsidian
Bug?) and only completely quitting and reopening would load the new
settings in the settings tab.